feat: The basic data structures Gradient, Keyframe and vector2Keyframe are moved out of the d3 library to the core library(基础数据结构Gradient、Keyframe、vector2Keyframe移出d3库到core库)
