+++ src/layaAir/Laya3D.ts
@@ -20,9 +20,8 @@ import { Scene3D } from "./laya/d3/core/scene/Scene3D";
 import { SkinnedMeshSprite3D } from "./laya/d3/core/SkinnedMeshSprite3D";
 import { Sprite3D } from "./laya/d3/core/Sprite3D";
 import { TrailMaterial } from "./laya/d3/core/trail/TrailMaterial";
-import { VertexTrail } from "./laya/d3/core/trail/VertexTrail";
 import { SubMeshInstanceBatch } from "./laya/d3/graphics/SubMeshInstanceBatch";
-import { VertexPositionTexture0 } from "./laya/d3/graphics/Vertex/VertexPositionTexture0";
+import { VertexPositionTexture } from "./laya/d3/graphics/Vertex/VertexPositionTexture";
 import { VertexShurikenParticleBillboard } from "./laya/d3/graphics/Vertex/VertexShurikenParticleBillboard";
 import { VertexShurikenParticleMesh } from "./laya/d3/graphics/Vertex/VertexShurikenParticleMesh";
 import { Mesh } from "./laya/d3/resource/models/Mesh";
@@ -41,7 +40,6 @@ import { Camera } from "./laya/d3/core/Camera";
 import { BaseCamera } from "./laya/d3/core/BaseCamera";
 import { ShuriKenParticle3DShaderDeclaration } from "./laya/d3/core/particleShuriKen/ShuriKenParticle3DShaderDeclaration";
 import { BaseRender } from "./laya/d3/core/render/BaseRender";
-import { TrailFilter } from "./laya/d3/core/trail/TrailFilter";
 import { DepthPass } from "./laya/d3/depthMap/DepthPass";
 import { RenderCapable } from "./laya/RenderEngine/RenderEnum/RenderCapable";
 import { BlitFrameBufferCMD } from "./laya/d3/core/render/command/BlitFrameBufferCMD";
@@ -135,8 +133,7 @@ export class Laya3D {
         VertexMesh.__init__();
         VertexShurikenParticleBillboard.__init__();
         VertexShurikenParticleMesh.__init__();
-        VertexPositionTexture0.__init__();
-        VertexTrail.__init__();
+        VertexPositionTexture.__init__();
         PixelLineVertex.__init__();
         SubMeshInstanceBatch.__init__();
         ShaderInit3D.__init__();
@@ -158,7 +155,6 @@ export class Laya3D {
         DepthPass.__init__();
         SkinnedMeshSprite3D.__init__();
         SimpleSkinnedMeshSprite3D.__init__();
-        TrailFilter.__init__();
         ShuriKenParticle3D.__init__();
         PostProcess.__init__();
         Scene3D.__init__();
