+++ src/layaAir/laya/d3/ModuleDef.ts
@@ -2,7 +2,6 @@ import { ClassUtils } from "../utils/ClassUtils";
 import { AnimationClip } from "./animation/AnimationClip";
 import { Sprite3D } from "./core/Sprite3D";
 import { Camera } from "./core/Camera";
-import { Gradient } from "./core/Gradient";
 import { LightSprite } from "./core/light/LightSprite";
 import { BlinnPhongMaterial } from "./core/material/BlinnPhongMaterial";
 import { EffectMaterial } from "./core/material/EffectMaterial";
@@ -68,7 +67,6 @@ import { AreaLightCom } from "./core/light/AreaLightCom";
 import { LODGroup, LODInfo } from "./component/LODGroup";
 import { PixelLineData } from "./core/pixelLine/PixelLineData";
 import { PixelLineRenderer } from "./core/pixelLine/PixelLineRenderer";
-import { FloatKeyframe } from "./core/FloatKeyframe";
 import { Animator } from "./component/Animator/Animator";
 import { AnimatorControllerLayer } from "./component/Animator/AnimatorControllerLayer";
 import { AnimatorState } from "./component/Animator/AnimatorState";
@@ -121,7 +119,7 @@ c("TrailFilter", TrailFilter);
 c("TrailRenderer", TrailRenderer);
 c("PixelLineRenderer", PixelLineRenderer);
 c("PixelLineData", PixelLineData);
-c("Gradient", Gradient);
+
 c("Transform3D", Transform3D);
 c("Lightmap", Lightmap);
 c("ReflectionProbe", ReflectionProbe);
@@ -134,7 +132,7 @@ c("Bounds", Bounds);
 c("BoundBox", BoundBox);
 c("TextureCube", TextureCube);
 c("Mesh", Mesh);
-c("FloatKeyframe", FloatKeyframe);
+
 c("RenderTexture", RenderTexture);
 c("Animator", Animator);
 c("AnimatorController", AnimatorController);
