--- src/layaAir/laya/d3/core/trail/TrailAlignment.ts
@@ -1,16 +1,0 @@
-/**
- * @en Enum for trail alignment options.
- * @zh 拖尾对齐方式枚举。
- */
-export enum TrailAlignment {
-    /**
-     * @en Align the trail to face the camera.
-     * @zh 使拖尾面向摄像机。
-     */
-    View,
-    /**
-     * @en Align the trail with the direction of the component
-     * @zh 使拖尾与组件的方向对齐。
-     */
-    TransformZ
-}
