+++ src/layaAir/laya/d3/core/trail/TrailMaterial.ts
@@ -4,6 +4,7 @@ import { RenderState } from "../../../RenderDriver/RenderModuleData/Design/Rende
 import { Shader3D } from "../../../RenderEngine/RenderShader/Shader3D";
 import { BaseTexture } from "../../../resource/BaseTexture";
 import { Material, MaterialRenderMode } from "../../../resource/Material";
+import { TrailShaderInit } from "./Trail/TrailShaderInit";
 
 /**
  * @en The `TrailMaterial` class is used to implement trail materials.
@@ -31,6 +32,7 @@ export class TrailMaterial extends Material {
 		TrailMaterial.MAINTEXTURE = Shader3D.propertyNameToID("u_MainTexture");
 		TrailMaterial.TINTCOLOR = Shader3D.propertyNameToID("u_MainColor");
 		TrailMaterial.TILINGOFFSET = Shader3D.propertyNameToID("u_TilingOffset");
+		TrailShaderInit.init();
 	}
 
 	/**
