+++ src/layaAir/laya/d3/graphics/Vertex/VertexPositionTexture.ts
@@ -1,4 +1,3 @@
-import { IVertex } from "../IVertex"
 import { VertexDeclaration } from "../../../RenderEngine/VertexDeclaration";
 import { Vector2 } from "../../../maths/Vector2";
 import { Vector3 } from "../../../maths/Vector3";
@@ -10,19 +9,19 @@ import { VertexMesh } from "../../../RenderEngine/RenderShader/VertexMesh";
  * @internal
  * <code>VertexPositionNormalTexture</code> 类用于创建位置、纹理顶点结构。
  */
-export class VertexPositionTexture0 implements IVertex {
+export class VertexPositionTexture {
 
 	private static _vertexDeclaration: VertexDeclaration;
 
 	static get vertexDeclaration(): VertexDeclaration {
-		return VertexPositionTexture0._vertexDeclaration;
+		return VertexPositionTexture._vertexDeclaration;
 	}
 
 	/**
 	 * @internal
 	 */
 	static __init__(): void {
-		VertexPositionTexture0._vertexDeclaration = new VertexDeclaration(20, [new VertexElement(0, VertexElementFormat.Vector3, VertexMesh.MESH_POSITION0),
+		VertexPositionTexture._vertexDeclaration = new VertexDeclaration(20, [new VertexElement(0, VertexElementFormat.Vector3, VertexMesh.MESH_POSITION0),
 		new VertexElement(12, VertexElementFormat.Vector2, VertexMesh.MESH_TEXTURECOORDINATE0)]);
 	}
 
@@ -38,7 +37,7 @@ export class VertexPositionTexture0 implements IVertex {
 	}
 
 	get vertexDeclaration(): VertexDeclaration {
-		return VertexPositionTexture0._vertexDeclaration;
+		return VertexPositionTexture._vertexDeclaration;
 	}
 
 	constructor(position: Vector3, textureCoordinate0: Vector2) {
