+++ src/layaAir/laya/d3/shader/ShaderInit3D.ts
@@ -32,7 +32,6 @@ import PBRLightingGLSL from "./lightingmode/PBRLighting.glsl";
 import { BlitScreenShaderInit } from "./postprocess/BlitScreenShaderInit";
 import { UnlitShaderInit } from "./unlit/UnlitShaderInit";
 import { BlinnPhongShaderInit } from "./blinnphong/BlinnPhongShaderInit";
-import { TrailShaderInit } from "./Trail/TrailShaderInit";
 import { PBRShaderLib } from "./pbr/PBRShaderLib";
 import { PBRStandardShaderInit } from "./pbr/PBRStandardShaderInit";
 import { SkyBoxShaderInit } from "./sky/SkyBoxShaderInit";
@@ -112,7 +111,6 @@ export class ShaderInit3D {
 		PBRStandardShaderInit.init();
 		//PBRShaderInit.init();
 		BlinnPhongShaderInit.init();
-		TrailShaderInit.init();
 		ParticleShuriKenShaderInit.init();
 		SkyBoxShaderInit.init();
 		SkyProceduralShaderInit.init();
