+++ src/layaAir/laya/display/RenderFeatureComman/Trail/TrailTextureMode.ts
@@ -1,0 +1,20 @@
+
+/**
+ * @en Trail Texture Mode
+ * @zh 拖尾纹理模式
+ */
+export class TrailTextureMode {
+    /**
+     * @en Stretch mode.
+     * @zh 拉伸模式。
+     */
+	static Stretch: number = 0;
+
+    /**
+     * @en Tile mode.
+     * @zh 平铺模式。
+     */
+	static Tile: number = 1;
+}
+
+
